Change OnMessage for new Conversation API.
The IBM Watson Conversation service has changed the object
that is returned from a call to the Conversation service.
Formerly, the object was a MessageResponse object that had
fields that could be accessed with dot syntax. Now, the object
is generic and the data returned must be converted, first to
fsdata and that is then converted to MessageResponse object.